Splitting the extension of g2o into multiple files
